backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
sqlalchemy
psycopg2-binary
anthropic
sentence-transformers
faiss-cpu
numpy
httpx
pypdf
fastembed
argon2-cffi
jinja2
